extract common env checker into spec helper
